Generate MIME that can be saved as draft #407
