+++ app/logic/Mail/SMTP/SMTPAccount.ts
@@ -1,13 +1,8 @@
-import { MailAccount, TLSSocketType, AuthMethod } from "../MailAccount";
+import { MailAccount, TLSSocketType, AuthMethod, NMUtils } from "../MailAccount";
 import type { EMail } from "../EMail";
-import type { PersonUID } from "../../Abstract/PersonUID";
-import { Attachment , ContentDisposition } from "../Attachment";
 import { appGlobal } from "../../app";
-import { getLocalStorage } from "../../../frontend/Util/LocalStorage";
-import { blobToBase64, assert } from "../../util/util";
+import { assert } from "../../util/util";
 import type { ArrayColl } from "svelte-collections";
-import type { default as Mail, Attachment as NMAttachment, Address as NMAddress } from "nodemailer/lib/mailer";
-type NMMail = Mail.Options;
 
 export class SMTPAccount extends MailAccount {
   readonly protocol: string = "smtp";
@@ -53,7 +48,7 @@ export class SMTPAccount extends MailAccount {
 
   async send(email: EMail): Promise<void> {
     try {
-      let mail = await SMTPAccount.getNMMail(email);
+      let mail = await NMUtils.getNMMail(email);
       let result = await appGlobal.remoteApp.sendMailNodemailer(
         this.getTransportOptions(), mail);
       email.sent = new Date();
@@ -68,32 +63,6 @@ export class SMTPAccount extends MailAccount {
     }
   }
 
-  static async getNMMail(email: EMail): Promise<NMMail> {
-    let doHTML = getLocalStorage("mail.send.format", "html").value == "html";
-    // <https://nodemailer.com/message/>
-    return {
-      subject: email.subject,
-      inReplyTo: email.inReplyTo,
-      from: SMTPAccount.getRecipient(email.from),
-      replyTo: email.replyTo ? SMTPAccount.getRecipient(email.replyTo) : null,
-      to: SMTPAccount.getRecipients(email.to),
-      cc: SMTPAccount.getRecipients(email.cc),
-      bcc: SMTPAccount.getRecipients(email.bcc),
-      text: email.text,
-      html: doHTML ? email.html : null,
-      attachDataUrls: true,
-      attachments: await SMTPAccount.getAttachments(email),
-      headers: email.headers.contentKeyValues(),
-      disableFileAccess: true,
-      disableUrlAccess: true,
-    };
-  }
-
-  static async getMIME(email: EMail): Promise<Uint8Array> {
-    let mail = await SMTPAccount.getNMMail(email);
-    return await appGlobal.remoteApp.getMIMENodemailer(mail);
-  }
-
   async verifyLogin(): Promise<void> {
     try {
       await appGlobal.remoteApp.verifyServerNodemailer(this.getTransportOptions());
@@ -104,29 +73,5 @@ export class SMTPAccount extends MailAccount {
       }
       throw ex;
     }
-  }
-
-  protected static getRecipients(recipients: ArrayColl<PersonUID>): NMAddress[] {
-    return recipients.contents.map(to => SMTPAccount.getRecipient(to));
-  }
-  protected static getRecipient(to: PersonUID): NMAddress {
-    // `${to.name} <${to.emailAddress}>` created by nodemailer, with quotes
-    return {
-      name: to.name,
-      address: to.emailAddress,
-    };
-  }
-  protected static async getAttachments(email: EMail): Promise<NMAttachment[]> {
-    return await Promise.all(email.attachments.contents.map((a) => SMTPAccount.getAttachment(a)));
-  }
-  protected static async getAttachment(a: Attachment): Promise<NMAttachment> {
-    return {
-      filename: a.filename,
-      content: await blobToBase64(a.content),
-      encoding: "base64",
-      contentType: a.mimeType,
-      contentDisposition: a.disposition == ContentDisposition.inline ? 'inline' : 'attachment',
-      cid: a.contentID,
-    };
   }
 }
